Seperate shared crafting behaviour
