+++ src/bpp_shared/inventory/interactions/crafting_table.cpp
@@ -1,0 +1,120 @@
+/*
+ * Copyright (c) 2026, jwaxy <jwaxy.is-a.dev>
+ *
+ * SPDX-License-Identifier: AGPL-3.0-only
+*/
+#include "crafting_table.h"
+#include "blocks.h"
+#include "inventory/interactions/crafting.h"
+#include "inventory/item_stack.h"
+#include "items.h"
+
+CraftingTableInventoryInteraction::CraftingTableInventoryInteraction(InventoryPlayer* pinv, WorldManager& worldMng,
+                                                                     Runtime& gameRuntime, Int3 craftingTablePos)
+    : CraftingInventoryInteraction(&sharedInventory, &craftInventory, pinv, gameRuntime, 3, 3), world(worldMng),
+      blockPosition(craftingTablePos) {
+	sharedInventory.owner = this;
+	mergeInventories();
+}
+
+CraftingTableInventoryInteraction::~CraftingTableInventoryInteraction() {
+	writeBack();
+
+	for (size_t i = 1; i <= gridWidth * gridHeight; i++) {
+		ItemStack& stack = craftInventory.slots[i];
+		if (stack.id == ITEM_INVALID)
+			continue;
+		playerInventory->mergeItemStackInInventory(stack, true, 9, 44);
+	}
+}
+
+void CraftingTableInventoryInteraction::writeBack() {
+	size_t slotCount = 0;
+	for (size_t i = 0; i < 10; i++)
+		craftInventory.slots[i] = sharedInventory.slots[slotCount++];
+	for (size_t i = 9; i < 45; i++)
+		playerInventory->slots[i] = sharedInventory.slots[slotCount++];
+}
+
+bool CraftingTableInventoryInteraction::canExist() {
+	return world.getBlockId(blockPosition) == BLOCK_CRAFTING_TABLE;
+}
+
+void CraftingTableInventoryInteraction::initSnapshot() {
+	snapshot.resize(size_t(sharedInventory.getSizeInventory()));
+	for (size_t i = 0; i < snapshot.size(); i++)
+		snapshot[i] = sharedInventory.slots[i];
+}
+
+std::vector<DeltaSlot> CraftingTableInventoryInteraction::tickDiff() {
+	std::vector<DeltaSlot> differences;
+	mergeInventories(); // make sure sharedInventory is current before diffing
+	for (size_t i = 0; i < snapshot.size(); i++) {
+		auto& snap = snapshot[i];
+		if (snap == sharedInventory.slots[i])
+			continue;
+		snap = sharedInventory.slots[i];
+		differences.push_back({ snap, int(i) });
+	}
+	return differences;
+}
+
+void CraftingTableInventoryInteraction::mergeInventories() {
+	size_t slotCount = 0;
+	for (size_t i = 0; i < 10; i++)
+		sharedInventory.slots[slotCount++] = craftInventory.slots[i];
+	for (size_t i = 9; i < 45; i++)
+		sharedInventory.slots[slotCount++] = playerInventory->slots[i];
+}
+
+void CraftingTableInventoryInteraction::updateResult() {
+	CraftingInventoryInteraction::updateResult();
+	mergeInventories();
+}
+
+void CraftingTableInventoryInteraction::shiftClickResult() {
+	ItemStack& result = craftInventory.slots[0];
+	if (result.id == ITEM_INVALID)
+		return;
+
+	ItemStack copy = result;
+	if (playerInventory->mergeItemStackInInventory(copy, true, 9, 44)) {
+		finishCraft();
+	} else {
+		craftInventory.slots[0] = copy.count == 0 ? ItemStack{} : copy;
+	}
+}
+
+void CraftingTableInventoryInteraction::shiftClickOther(int slot) {
+	auto stack = sharedInventory.getStackInSlot(slot);
+	if (!stack)
+		return;
+
+	ItemStack copy = *stack;
+
+	if (slot < 10) {
+		// Grid -> inventory
+		// Try the main inventory then the hotbar
+		bool success = playerInventory->mergeItemStackInInventory(copy, false, 9, 35);
+		if (!success)
+			playerInventory->mergeItemStackInInventory(copy, false, 36, 44);
+	} else {
+		// We can't shift click into the crafting grid itself, so just try the other area of the inventory
+		// We shift clicked in the inventory
+		if (slot > 9 && slot < 37) {
+			playerInventory->mergeItemStackInInventory(copy, false, 36, 44);
+		} else {
+			playerInventory->mergeItemStackInInventory(copy, false, 9, 35);
+		}
+	}
+
+	// Update the source in the real inventory before re-merging
+	if (slot < 10) {
+		craftInventory.slots[slot] = copy.count == 0 ? ItemStack{} : copy;
+	} else {
+		playerInventory->slots[slot - 1] = copy.count == 0 ? ItemStack{} : copy;
+	}
+
+	// Re-sync sharedInventory from the real inventories
+	mergeInventories();
+}
